+++ src/backend/mailbox_manager/api/client/viewsets.py
@@ -2,6 +2,7 @@
 
 from django.db.models import Q, Subquery
 from django.http import Http404
+from django.shortcuts import get_object_or_404
 
 from rest_framework import exceptions, filters, mixins, status, viewsets
 from rest_framework.decorators import action
@@ -9,11 +10,11 @@ from rest_framework.response import Response
 
 from core import models as core_models
 from core.api.client.serializers import UserSerializer
+from core.exceptions import EmailAlreadyKnownException
 
 from mailbox_manager import enums, models
 from mailbox_manager.api import permissions
 from mailbox_manager.api.client import serializers
-from mailbox_manager.exceptions import EmailAlreadyKnownException
 from mailbox_manager.utils.dimail import DimailAPIClient
 
 
@@ -419,7 +420,7 @@ class MailDomainInvitationViewset(
         try:
             return super().create(request, *args, **kwargs)
         except EmailAlreadyKnownException as exc:
-            user = models.User.objects.get(email=email)
+            user = models.User.objects.get(email__iexact=email)
 
             models.MailDomainAccess.objects.create(
                 user=user,
@@ -427,6 +428,14 @@ class MailDomainInvitationViewset(
                 role=request.data["role"],
             )
             raise exc
+
+    @action(detail=True, methods=["post"])
+    def refresh(self, request, *args, **kwargs):  # pylint: disable=unused-argument
+        """Enable mailbox. Send a request to dimail and change status in our DB"""
+        invitation = get_object_or_404(models.MailDomainInvitation, id=kwargs["id"])
+        invitation.refresh()
+        invitation.email_invitation()
+        return Response(serializers.MailDomainInvitationSerializer(invitation).data)
 
 
 class AliasViewSet(