+++ CHANGELOG.md
@@ -8,9 +8,17 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.23.1] - 2026-02-16
+
+- ✨(invitations) refresh expired invitations
+
+## [1.23.0] - 2026-02-12
+
 ### Added
 
 - ✨(datagrid) add sort to mailboxes list + mail domain list
+- ✨(demo) add aliases to demo #1050
+- ✨(front) add icon to button to configure a domain
 - ✨(invitations) allow delete invitations mails domains access by an admin
 - ✨(front) delete invitations mails domains access
 - ✨(front) add show invitations mails domains access #1040
@@ -22,6 +30,7 @@ and this project adheres to
 
 ### Changed
 
+- 🚸(email) we should ignore case when looking for existing emails #1056
 - 🏗️(core) migrate from pip to uv
 - ✨(front) add show invitations mails domains access #1040
 
@@ -503,7 +512,9 @@ and this project adheres to
 - ✨(domains) create and manage domains on admin + API
 - ✨(domains) mailbox creation + link to email provisioning API
 
-[unreleased]: https://github.com/suitenumerique/people/compare/v1.22.2...main
+[unreleased]: https://github.com/suitenumerique/people/compare/v1.23.1...main
+[1.23.1]: https://github.com/suitenumerique/people/releases/v1.23.1
+[1.23.0]: https://github.com/suitenumerique/people/releases/v1.23.0
 [1.22.2]: https://github.com/suitenumerique/people/releases/v1.22.2
 [1.22.1]: https://github.com/suitenumerique/people/releases/v1.22.1
 [1.22.0]: https://github.com/suitenumerique/people/releases/v1.22.0